Summary and dada2

+++ make_otu_table.sh
@@ -22,24 +22,29 @@ then
 make_otu_table.py -i $1 -t $2 -c $3 -of $outlocation -cluster_id ${9} -cluster_size "${10}"
 fi
 
+usearch10.0.240 -otutab_stats $outlocation"/otutab.txt" -output $outlocation/"report.txt" 2>&1
+echo "Otu table summary" >> $outlocation"/log.log"
+echo "============================================================" >> $outlocation"/log.log"
+cat $outlocation/"report.txt" >> $outlocation"/log.log" 
+
 #output files
-if [ $4 ]
+if [ $4 ] && [ -f $outlocation"/all_output.zip" ]
 then
     mv $outlocation"/all_output.zip" $4
 fi
-if [ $5 ]
+if [ $5 ] && [ -f $outlocation"/log.log" ]
 then
     mv $outlocation"/log.log" $5
 fi
-if [ $6 ]
+if [ $6 ] && [ -f $outlocation"/otu_sequences.fa" ]
 then
     mv $outlocation"/otu_sequences.fa" $6
 fi
-if [ $7 ]
+if [ $7 ] && [ -f $outlocation"/otutab.txt" ]
 then
     mv $outlocation"/otutab.txt" $7
 fi
-if [ $8 ]
+if [ $8 ] && [ -f $outlocation"/bioom.json" ]
 then
     mv $outlocation"/bioom.json" $8
 fi
